// frontend/lib/api/generated/person/person.ts
export type getApiPersonResponse200TextPlain = {
  data: PersonPagedResult
  status: 200
}

export type getApiPersonResponse200ApplicationJson = {
  data: PersonPagedResult
  status: 200
}

export type getApiPersonResponse200TextJson = {
  data: PersonPagedResult
  status: 200
}

export type getApiPersonResponse400TextPlain = {
  data: ProblemDetails
  status: 400
}

export type getApiPersonResponse400ApplicationJson = {
  data: ProblemDetails
  status: 400
}

export type getApiPersonResponse400TextJson = {
  data: ProblemDetails
  status: 400
}

export type getApiPersonResponse500 = {
  data: void
  status: 500
}

export type getApiPersonResponseSuccess = (getApiPersonResponse200TextPlain | getApiPersonResponse200ApplicationJson | getApiPersonResponse200TextJson) & {
  headers: Headers;
};