// frontend/lib/api/generated/model/socialSecurityRegistration.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Prevly.Api
 * OpenAPI spec version: 1.0
 */
import type { SocialSecurityRegistrationStatus } from './socialSecurityRegistrationStatus';

export interface SocialSecurityRegistration {
  /** @nullable */
  id?: string | null;
  /** @nullable */
  number?: string | null;
  /** @nullable */
  firstContributionDate?: string | null;
  /** @nullable */
  lastContributionDate?: string | null;
  contributionYears?: number;
  createdAt?: string;
  /** @nullable */
  personId?: string | null;
  status?: SocialSecurityRegistrationStatus;
  /** @nullable */
  ownershipCheckedAt?: string | null;
  /** @nullable */
  ownershipOwnerName?: string | null;
  /** @nullable */
  lastProcessingError?: string | null;
}